Truncate segment files of AO relations before unlink (#675)
Problem description:
Segment files of the AO tables were not truncated before unlink. As a result,
after relation drop, if there were file descriptors of segment files, not closed
by some backend process, disk space was not returned to the OS.

Root cause:
In mdunlinkfork, mdunlink_ao is called for an AO relation instead of truncating
all segment files. And mdunlink_ao doesn't perform truncation of segment files.

Fix:
Truncation of segment files for AO tables was added into mdunlink_ao_perFile,
which is called from mdunlink_ao for each segment file. For that purpose, static
function do_truncate was changed to global. Plus aomd unit test was updated.
Reason - unit test calls mdunlink_ao with filenames of not-existing files. Unit
test replaces unlink, so it can handle such filenames. But now
mdunlink_ao_perFile returns before invoking unlink, because do_truncate fails to
do truncation of these files. To fix it, do_truncate in aomd unit test was
replaced with a stub during the test.

Changes from original commit:
1. Added RESET for `default_tablespace` to previous test.

(cherry picked from commit 6f0a2c14350642400b695e9d87eb809ba23d6a98)

+++ src/backend/access/appendonly/aomd.c
@@ -42,6 +42,7 @@
 #include "common/relpath.h"
 #include "pgstat.h"
 #include "storage/bufmgr.h"
+#include "storage/md.h"
 #include "storage/sync.h"
 #include "utils/faultinjector.h"
 #include "utils/guc.h"
@@ -369,6 +370,13 @@ mdunlink_ao_perFile(const int segno, void *ctx)
 	INIT_FILETAG(tag, unlinkFiles->rnode, MAIN_FORKNUM, segno,
 				 SYNC_HANDLER_AO);
 	RegisterSyncRequest(&tag, SYNC_FORGET_REQUEST, true);
+
+	/*
+	 * unlink is not enough to return disk space to the OS immediately, because
+	 * the file can be still opened by other process
+	 */
+	if (do_truncate(segPath) < 0 && errno == ENOENT)
+		return false;
 
 	/* Next unlink the file */
 	if (unlink(segPath) != 0)
